BC H of C Debates
NOTE - Martin - I have also uploaded the original PDF of the applicable
pages. The OCR for this file seems pretty incredible, and I will leave
it to you whether it is better to 1) use our normal workflow with JPGs,
or 2) convert the existing OCRed PDF pages into single-page html
documents.