PEI H of C debates upload
Martin - please prepare these files before doing the rest of the
Maritimes - the assigned transcriber is very eager to be part of the
beta test. Please note:
1) These files are all very good OCR, so this might be one of the cases
where it is easier to extract the text, rather than OCR it.
2) a few of the volumes show French and English side by side, while
others only have English. We will obviously only want to include the
English texts for the time being. I can delete the non-applicable
content from the HTML files before it goes to the transcriber if that
is easier for you.